try to get tests running
